pyproject.toml
[project]
name = "rvongunten-office-automation"
version = "0.1.0"
description = "Telegram Bot mit AI-Dokumentenerkennung und RPA für Infoniqa ONE Start"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=21.0",
    "anthropic>=0.40.0",
    "pymupdf>=1.24.0",
    "sqlalchemy>=2.0",
    "alembic>=1.14.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
rpa = [
    "pywinauto>=0.6.8",
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
